Asserting that '' is not a ItemBarcode
This is an interesting question, is '' an ItemBarcode?

If "" is a valid ItemBarCode then an ActiveRecord::InvalidRecord
exception is thrown on this line:

https://github.com/ndlib/annex-ims/blob/5d5e303e5fb6d4faa55157f593fc9feda78ff863/app/services/get_item_from_barcode.rb#L35

If "" is not a valid ItemBarCode then a StandardError exception is
thrown.

https://github.com/ndlib/annex-ims/blob/5d5e303e5fb6d4faa55157f593fc9feda78ff863/app/services/get_item_from_barcode.rb#L21

It is important to decide what is the correct answer as we the
handling of [AIMS-271](https://jira.library.nd.edu/browse/AIMS-271)
will vary