Updates for users page AIMS-335
Admins can now update the type of user for existing users in the database via the users page, which now has select options for user types and update button. Every time the user type is changed, there will be a notification on top of the page that will give details of the update. Also the current user is disabled to change own user type due to a safety requirement.