AIMS-354: Added validation for count of items in the tray
When a tray is full the user clicks "Done".  At that point the user is prompted to "Enter the number of items in the tray."
If the count matches, the user is taken to the Stocking page to scan the next Tray.
If the count does NOT match, the user receives an error message and is prompted to recount, and enter the count again.