AIMS-355: The second error in Tray Count directs to set the tray aside
Users are given a second chance to enter the Tray Count.
If the numbers match on the 2nd entry, then they are rerouted to the Tray Stocking page where they can scan the next tray.
If the numbers do NOT match, then the user is directed to set the Tray aside to be reviewed.
Once they click "OK" to this prompt, then they are rerouted to the Tray Stocking page where they can scan the next tray.